+++ Projects/UOContent/Engines/ConPVP/Trophy.cs
@@ -10,6 +10,22 @@ public enum TrophyRank
     Gold
 }
 
+public static class TrophyRankExtensions
+{
+    /// <summary>
+    /// Returns a static lowercase string for the given <see cref="TrophyRank"/> value.
+    /// Avoids the two-allocation <c>rank.ToString().ToLower()</c> pattern at interpolation sites.
+    /// </summary>
+    public static string LowerName(this TrophyRank rank) =>
+        rank switch
+        {
+            TrophyRank.Bronze => "bronze",
+            TrophyRank.Silver => "silver",
+            TrophyRank.Gold   => "gold",
+            _                 => rank.ToString().ToLowerInvariant()
+        };
+}
+
 [Flippable(5020, 4647)]
 [SerializationGenerator(2, false)]
 public partial class Trophy : Item
@@ -122,3 +138,4 @@ public partial class Trophy : Item
         };
     }
 }
+
